src/test/java/com/jpaExample/jpaExample1/service/MemberServiceTest.java
package com.jpaExample.jpaExample1.service;

import com.jpaExample.jpaExample1.domain.member.Member;
import com.jpaExample.jpaExample1.repository.MemberRepository;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.context.web.WebAppConfiguration;
import org.springframework.transaction.annotation.Transactional;import static org.junit.Assert.assertEquals;

@RunWith(SpringRunner.class)
@SpringBootTest
@Transactional
public class MemberServiceTest {

    @Autowired
    MemberService memberService;

    @Autowired
    MemberRepository memberRepository;

    @Test
    public void joinTest() throws Exception {
        //Given
        Member member = new Member();
//        member.("kim");
        //When
        Long saveId = memberService.join(member);
        //Then
        assertEquals(member, memberRepository.findByOne(saveId));
    }

    @Test(expected = IllegalStateException.class)
    public void dupulicateTest() throws Exception {
        //Given
        Member member1 = new Member();
        member1.setName("kim");
        Member member2 = new Member();
        member2.setName("kim");
        //When
        memberService.join(member1);
        memberService.join(member2); //예외가 발생해야 한다.
        //Then fail("예외가 발생해야 한다.");
    }
}